corrected an error with high porta-speeds. added digi.txt, digibooster description.
git-svn-id: https://pimpmobile.svn.sourceforge.net/svnroot/pimpmobile/trunk@75 3d5ecaf0-f903-0410-b953-c2c1a4d75763

+++ src/internal.h
@@ -41,7 +41,7 @@ typedef struct
 	s32 period;
 	s32 final_period; /* signed so we can check for underfow */
 	s32 porta_target;
-	s8  porta_speed;
+	u16 porta_speed;
 	s8  volume_slide_speed;
 
 	s8  volume;
